fix: adjust price typing in register service schema

+++ src/Handlers/Services/Register/RegisterServiceSchema.ts
@@ -2,7 +2,7 @@ import * as zod from "zod";
 
 export const RegisterServiceSchema = zod.object({
   name: zod.string().min(3).max(100),
-  price: zod.number().positive().int(),
+  price: zod.number().positive(),
 });
 
 export type RegisterServiceCommand = zod.infer<typeof RegisterServiceSchema>;
